--- src/main/kotlin/dev/usbharu/hideout/core/service/auth/RecaptchaResult.kt
@@ -1,9 +1,0 @@
-package dev.usbharu.hideout.core.service.auth
-
-data class RecaptchaResult(
-    val success: Boolean,
-    val challenge_ts: String,
-    val hostname: String,
-    val score: Float,
-    val action: String
-)
