--- src/main/kotlin/dev/usbharu/hideout/core/service/auth/RegisterAccountDto.kt
@@ -1,7 +1,0 @@
-package dev.usbharu.hideout.core.service.auth
-
-data class RegisterAccountDto(
-    val username:String,
-    val password:String,
-    val recaptchaResponse:String
-)
